% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechatRebort.R
\name{article}
\alias{article}
\title{Create an article for the news type messages}
\usage{
article(title, discription, url, picurl = NULL)
}
\arguments{
\item{title}{The new title}

\item{discription}{The discription of the news}

\item{url}{a url for the message}

\item{picurl}{a picture url}
}
\value{
an article list
}
\description{
Create an article for the news type messages
}
\examples{
article <- article(title = "中秋节礼品领取",
discription =  "今年中秋节公司有豪礼相送",
url = "www.qq.com",
picurl = "some url")
}
