% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechatRebort.R
\name{sendImages}
\alias{sendImages}
\title{Send Images Using Qiye Wechat}
\usage{
sendImages(webhook, image_filepath)
}
\arguments{
\item{webhook}{an api link from Qiye wechat}

\item{image_filepath}{the filepath of the image}
}
\value{
a respose type
}
\description{
Send Images Using Qiye Wechat
}
\note{
webhook <-  "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=XXXXXXXXXXX"
sendImages(webhook,"full file path")
}
