% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechatRebort.R
\name{sendMarkdown}
\alias{sendMarkdown}
\title{Send Makrdown Message using Qiye Wechat}
\usage{
sendMarkdown(webhook, markdown)
}
\arguments{
\item{webhook}{an api link from Qiye wechat}

\item{markdown}{a string with markdown}
}
\value{
a respose type
}
\description{
Send Makrdown Message using Qiye Wechat
}
\note{
see https://work.weixin.qq.com/api/doc/90000/90136/91770 for more infomation about
supported markdown languages
}
\examples{
webhook <-  "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=XXXXXXXXXXX"
markdown <- paste0('实时新增用户反馈<font color=\"warning\">132例</font>，请相关同事注意。
> 类型:<font color=\"comment\">用户反馈</font>
> 普通用户反馈:<font color=\"comment\">117例</font>
> VIP用户反馈:<font color=\"comment\">15例</font>')
response <- sendMarkdown(webhook, markdown)
}
