% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechatRebort.R
\name{sendNews}
\alias{sendNews}
\title{Send articles using Qiye Wechat}
\usage{
sendNews(webhook, articles)
}
\arguments{
\item{webhook}{an api link from Qiye wechat}

\item{articles}{a list of articles to send}
}
\value{
a response type
}
\description{
Send articles using Qiye Wechat
}
\examples{
webhook <-  "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=XXXXXXXXXXX"
article_1 <- article(title = "中秋节礼品领取",
discription =  "今年中秋节公司有豪礼相送",
url = "www.qq.com")
article_2 <- article(title = "中秋节礼品领取",
discription =  "今年中秋节公司有豪礼相送",
url = "www.qq.com")
articles <- list(article_1,article_2)
sendNews(webhook = webhook, articles = articles)
}
