% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wechatRebort.R
\name{sendText}
\alias{sendText}
\title{send a test message using qiye Wechat}
\usage{
sendText(webhook, content, mentioned_list = NULL)
}
\arguments{
\item{webhook}{an api link from Qiye wechat}

\item{content}{a string}

\item{mentioned_list}{a list type}
}
\value{
a respose type
}
\description{
send a test message using qiye Wechat
}
\examples{
webhook <- "https://qyapi.weixin.qq.com/cgi-bin/webhook/send?key=XXXXXXXXXXX"
response <- sendText(webhook,"hello world!")

response <- sendText(webhook,"hello world!",list("someone","@all"))

}
